[Integration Test] spilt the controller and test.
