+++ restlight-integration-test/springmvc-test/src/main/java/io/esastack/restlight/integration/cases/controller/RestAnnotationController.java
@@ -18,7 +18,6 @@ import io.esastack.restlight.core.annotation.RequestBean;
 import io.esastack.restlight.integration.cases.annotation.CustomRequestBean;
 import io.esastack.restlight.integration.cases.annotation.CustomRequestBody;
 import io.esastack.restlight.integration.cases.annotation.CustomResponseBody;
-import io.esastack.restlight.integration.cases.exception.CustomException;
 import io.esastack.restlight.integration.entity.UserData;
 import org.springframework.http.HttpStatus;
 import org.springframework.util.MultiValueMap;
@@ -96,75 +95,25 @@ public class RestAnnotationController {
                 .name(name).build();
     }
 
-    @GetMapping("get/reponsestatus")
+    @GetMapping("get/responsestatus")
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
     public UserData responseStatus() {
         return UserData.Builder.aRestResult().build();
     }
 
-    @GetMapping("get/custombean")
+    @GetMapping("get/custom/bean")
     public UserData customBean(@CustomRequestBean UserData user) {
         return user;
     }
 
-    @PostMapping("post/custombody")
+    @PostMapping("post/custom/body")
     public UserData customBody(@CustomRequestBody UserData user) {
         return user;
     }
 
-    @GetMapping("get/customresponsebody")
+    @GetMapping("get/custom/responsebody")
     @CustomResponseBody
     public String customResponseBody(@RequestParam String name) {
         return name;
-    }
-
-    @GetMapping("get/filter")
-    public UserData filter(@RequestHeader String name) {
-        return UserData.Builder.aRestResult()
-                .name(name).build();
-    }
-
-    @GetMapping("get/exception")
-    public void exception() {
-        throw new RuntimeException("Forbidden");
-    }
-
-    @GetMapping("get/customexception")
-    public void customException() {
-        throw new CustomException("Custom");
-    }
-
-    @GetMapping("get/paramadvicefactory")
-    public UserData customParamAdviceByFactory(@RequestParam String name) {
-        return UserData.Builder.aRestResult()
-                .name(name).build();
-    }
-
-    @GetMapping("get/paramadviceadaptor")
-    public UserData customParamAdviceByAdaptor(@RequestParam String name) {
-        return UserData.Builder.aRestResult()
-                .name(name).build();
-    }
-
-    @PostMapping("post/entityadvicefactory")
-    public UserData customEntityAdviceByFactory(@RequestBody UserData user) {
-        return user;
-    }
-
-    @PostMapping("post/entityadviceadaptor")
-    public UserData customEntityAdviceByAdaptor(@RequestBody UserData user) {
-        return user;
-    }
-
-    @GetMapping("get/response/entityadvicefactory")
-    public UserData customResponseEntityAdviceByFactory(@RequestParam String name) {
-        return UserData.Builder.aRestResult()
-                .name(name).build();
-    }
-
-    @GetMapping("get/response/entityadviceadaptor")
-    public UserData customResponseEntityAdviceByAdaptor(@RequestParam String name) {
-        return UserData.Builder.aRestResult()
-                .name(name).build();
     }
 }
