+++ restlight-integration-test/springmvc-test/src/main/java/io/esastack/restlight/integration/test/RestAnnotationTest.java
@@ -108,14 +108,14 @@ public class RestAnnotationTest extends BaseIntegrationTest {
 
     @Test
     public void testResponseStatus() throws Exception {
-        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/reponsestatus").execute()
+        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/responsestatus").execute()
                 .toCompletableFuture().get();
         Assert.assertEquals(HttpStatus.INTERNAL_SERVER_ERROR.code(), response.status());
     }
 
     @Test
     public void testCustomBean() throws Exception {
-        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/custombean")
+        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/custom/bean")
                 .addParam("name", "test").execute()
                 .toCompletableFuture().get();
         UserData userData = response.bodyToEntity(UserData.class);
@@ -127,7 +127,7 @@ public class RestAnnotationTest extends BaseIntegrationTest {
         UserData user = UserData.Builder.aRestResult()
                 .name("test").age(10).birthDay(new Date())
                 .weight(BigDecimal.valueOf(123.01)).build();
-        RestResponseBase response = restClient.post(domain + "/rest/annotation/post/custombody").entity(user).execute()
+        RestResponseBase response = restClient.post(domain + "/rest/annotation/post/custom/body").entity(user).execute()
                 .toCompletableFuture().get();
         UserData userResult = response.bodyToEntity(UserData.class);
         Assert.assertEquals(user.getName(), userResult.getName());
@@ -135,87 +135,10 @@ public class RestAnnotationTest extends BaseIntegrationTest {
 
     @Test
     public void testCustomResponseBody() throws Exception {
-        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/customresponsebody")
+        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/custom/responsebody")
                 .addParam("name", "test").execute()
                 .toCompletableFuture().get();
         UserData user = response.bodyToEntity(UserData.class);
         Assert.assertEquals("test", user.getName());
-    }
-
-    @Test
-    public void testFilter() throws Exception {
-        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/filter")
-                .addParam("name", "test").execute()
-                .toCompletableFuture().get();
-        UserData user = response.bodyToEntity(UserData.class);
-        Assert.assertEquals("test", user.getName());
-    }
-
-    @Test
-    public void testException() throws Exception {
-        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/exception")
-                .execute().toCompletableFuture().get();
-        Assert.assertEquals(HttpStatus.FORBIDDEN.code(), response.status());
-        Assert.assertEquals("Forbidden", response.bodyToEntity(String.class));
-    }
-
-    @Test
-    public void testCustomException() throws Exception {
-        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/customexception")
-                .execute().toCompletableFuture().get();
-        Assert.assertEquals(HttpStatus.UNAUTHORIZED.code(), response.status());
-        Assert.assertEquals("Custom", response.bodyToEntity(String.class));
-    }
-
-    @Test
-    public void testCustomParamAdviceByFactory() throws Exception {
-        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/paramadvicefactory")
-                .addParam("name", "test").execute().toCompletableFuture().get();
-        UserData user = response.bodyToEntity(UserData.class);
-        Assert.assertEquals("test-advice-factory", user.getName());
-    }
-
-    @Test
-    public void testCustomParamAdviceByAdaptor() throws Exception {
-        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/paramadviceadaptor")
-                .addParam("name", "test").execute().toCompletableFuture().get();
-        UserData user = response.bodyToEntity(UserData.class);
-        Assert.assertEquals("test-advice-adaptor", user.getName());
-    }
-
-    @Test
-    public void testCustomEntityAdviceByFactory() throws Exception {
-        UserData user = UserData.Builder.aRestResult()
-                .name("test").build();
-        RestResponseBase response = restClient.post(domain + "/rest/annotation/post/entityadvicefactory")
-                .entity(user).execute().toCompletableFuture().get();
-        UserData userData = response.bodyToEntity(UserData.class);
-        Assert.assertEquals(user.getName() + "-advice-factory", userData.getName());
-    }
-
-    @Test
-    public void testCustomEntityAdviceByAdaptor() throws Exception {
-        UserData user = UserData.Builder.aRestResult()
-                .name("test").build();
-        RestResponseBase response = restClient.post(domain + "/rest/annotation/post/entityadviceadaptor")
-                .entity(user).execute().toCompletableFuture().get();
-        UserData userData = response.bodyToEntity(UserData.class);
-        Assert.assertEquals(user.getName() + "-advice-adaptor", userData.getName());
-    }
-
-    @Test
-    public void testCustomResponseEntityAdviceByFactory() throws Exception {
-        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/response/entityadvicefactory")
-                .addParam("name", "test").execute().toCompletableFuture().get();
-        UserData user = response.bodyToEntity(UserData.class);
-        Assert.assertEquals("test-advice-factory", user.getName());
-    }
-
-    @Test
-    public void testCustomResponseEntityAdviceByAdaptor() throws Exception {
-        RestResponseBase response = restClient.get(domain + "/rest/annotation/get/response/entityadviceadaptor")
-                .addParam("name", "test").execute().toCompletableFuture().get();
-        UserData user = response.bodyToEntity(UserData.class);
-        Assert.assertEquals("test-advice-adaptor", user.getName());
     }
 }
